+++ pkg/knative/commands/handler_invoke_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package knative_commands_test
+package commands_test
 
 import (
 	"fmt"
@@ -26,8 +26,7 @@ import (
 	"github.com/knative/pkg/apis"
 	duckv1beta1 "github.com/knative/pkg/apis/duck/v1beta1"
 	"github.com/projectriff/cli/pkg/cli"
-	knativecommands "github.com/projectriff/cli/pkg/knative/commands"
-
+	"github.com/projectriff/cli/pkg/knative/commands"
 	rifftesting "github.com/projectriff/cli/pkg/testing"
 	requestv1alpha1 "github.com/projectriff/system/pkg/apis/knative/v1alpha1"
 	corev1 "k8s.io/api/core/v1"
@@ -39,21 +38,21 @@ func TestHandlerInvokeOptions(t *testing.T) {
 	table := rifftesting.OptionsTable{
 		{
 			Name: "invalid resource",
-			Options: &knativecommands.HandlerInvokeOptions{
+			Options: &commands.HandlerInvokeOptions{
 				ResourceOptions: rifftesting.InvalidResourceOptions,
 			},
 			ExpectFieldError: rifftesting.InvalidResourceOptionsFieldError,
 		},
 		{
 			Name: "valid resource",
-			Options: &knativecommands.HandlerInvokeOptions{
+			Options: &commands.HandlerInvokeOptions{
 				ResourceOptions: rifftesting.ValidResourceOptions,
 			},
 			ShouldValidate: true,
 		},
 		{
 			Name: "json content type",
-			Options: &knativecommands.HandlerInvokeOptions{
+			Options: &commands.HandlerInvokeOptions{
 				ResourceOptions: rifftesting.ValidResourceOptions,
 				ContentTypeJSON: true,
 			},
@@ -61,7 +60,7 @@ func TestHandlerInvokeOptions(t *testing.T) {
 		},
 		{
 			Name: "text content type",
-			Options: &knativecommands.HandlerInvokeOptions{
+			Options: &commands.HandlerInvokeOptions{
 				ResourceOptions: rifftesting.ValidResourceOptions,
 				ContentTypeText: true,
 			},
@@ -69,7 +68,7 @@ func TestHandlerInvokeOptions(t *testing.T) {
 		},
 		{
 			Name: "multiple content types",
-			Options: &knativecommands.HandlerInvokeOptions{
+			Options: &commands.HandlerInvokeOptions{
 				ResourceOptions: rifftesting.ValidResourceOptions,
 				ContentTypeJSON: true,
 				ContentTypeText: true,
@@ -371,7 +370,7 @@ Command executed: curl localhost -H 'Host: test-handler.example.com'
 		},
 	}
 
-	table.Run(t, knativecommands.NewHandlerInvokeCommand)
+	table.Run(t, commands.NewHandlerInvokeCommand)
 }
 
 func TestHelperProcess_HandlerInvoke(t *testing.T) {
